put type definitions to separate file
